source/requirements.txt
numpy
mxnet
opencv-python
websocket-client
easydict